+++ libnavui-app/src/test/java/com/mapbox/navigation/ui/app/internal/controller/AudioGuidanceStateControllerTest.kt
@@ -1,7 +1,5 @@
 package com.mapbox.navigation.ui.app.internal.controller
 
-import com.mapbox.api.directions.v5.models.VoiceInstructions
-import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.MapboxNavigation
 import com.mapbox.navigation.core.lifecycle.MapboxNavigationApp
 import com.mapbox.navigation.testing.MainCoroutineRule
@@ -10,8 +8,8 @@ import com.mapbox.navigation.ui.app.internal.audioguidance.AudioAction
 import com.mapbox.navigation.ui.app.internal.audioguidance.AudioGuidanceState
 import com.mapbox.navigation.ui.app.internal.navigation.NavigationState
 import com.mapbox.navigation.ui.app.testing.TestStore
-import com.mapbox.navigation.ui.voice.internal.MapboxAudioGuidance
-import com.mapbox.navigation.ui.voice.model.SpeechAnnouncement
+import com.mapbox.navigation.ui.voice.api.MapboxAudioGuidance
+import com.mapbox.navigation.ui.voice.api.MapboxAudioGuidanceState
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.mockkObject
@@ -27,13 +25,13 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
-@OptIn(ExperimentalPreviewMapboxNavigationAPI::class, ExperimentalCoroutinesApi::class)
+@OptIn(ExperimentalCoroutinesApi::class)
 class AudioGuidanceStateControllerTest {
 
     @get:Rule
     var coroutineRule = MainCoroutineRule()
 
-    private lateinit var audioGuidanceState: MutableStateFlow<MapboxAudioGuidance.State>
+    private lateinit var audioGuidanceState: MutableStateFlow<MapboxAudioGuidanceState>
     private lateinit var mockAudioGuidance: MapboxAudioGuidance
     private lateinit var testStore: TestStore
 
@@ -42,7 +40,7 @@ class AudioGuidanceStateControllerTest {
         mockkObject(MapboxNavigationApp)
 
         audioGuidanceState = MutableStateFlow(
-            TestAudioGuidanceState(isMuted = false)
+            mockk { every { isMuted } returns false }
         )
         mockAudioGuidance = mockk(relaxed = true) {
             every { stateFlow() } returns audioGuidanceState
@@ -68,7 +66,7 @@ class AudioGuidanceStateControllerTest {
                     audio = AudioGuidanceState(isMuted = false)
                 )
             )
-            audioGuidanceState.value = TestAudioGuidanceState(isMuted = true)
+            audioGuidanceState.value = mockk { every { isMuted } returns true }
 
             sut.onAttached(mockMapboxNavigation())
             assertTrue(testStore.state.value.audio.isMuted)
@@ -83,7 +81,7 @@ class AudioGuidanceStateControllerTest {
                     audio = AudioGuidanceState(isMuted = false)
                 )
             )
-            audioGuidanceState.value = TestAudioGuidanceState(isMuted = false)
+            audioGuidanceState.value = mockk { every { isMuted } returns false }
 
             sut.onAttached(mockMapboxNavigation())
             testStore.setState(
@@ -118,11 +116,4 @@ class AudioGuidanceStateControllerTest {
         every { MapboxNavigationApp.current() } returns mapboxNavigation
         return mapboxNavigation
     }
-
-    private data class TestAudioGuidanceState(
-        override val isMuted: Boolean,
-        override val isPlayable: Boolean = false,
-        override val voiceInstructions: VoiceInstructions? = null,
-        override val speechAnnouncement: SpeechAnnouncement? = null
-    ) : MapboxAudioGuidance.State
 }