+++ libnavui-voice/src/main/java/com/mapbox/navigation/ui/voice/installer/ComponentInstaller.kt
@@ -1,12 +1,10 @@
 package com.mapbox.navigation.ui.voice.installer
 
-import android.content.Context
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.lifecycle.MapboxNavigationApp
 import com.mapbox.navigation.ui.base.installer.ComponentInstaller
 import com.mapbox.navigation.ui.base.installer.Installation
-import com.mapbox.navigation.ui.voice.internal.MapboxAudioGuidance
-import com.mapbox.navigation.ui.voice.internal.impl.MapboxAudioGuidanceImpl
+import com.mapbox.navigation.ui.voice.api.MapboxAudioGuidance
 import com.mapbox.navigation.ui.voice.internal.ui.AudioGuidanceButtonComponent
 import com.mapbox.navigation.ui.voice.view.MapboxAudioGuidanceButton
 
@@ -19,14 +17,13 @@ import com.mapbox.navigation.ui.voice.view.MapboxAudioGuidanceButton
  */
 @ExperimentalPreviewMapboxNavigationAPI
 fun ComponentInstaller.audioGuidanceButton(button: MapboxAudioGuidanceButton): Installation {
-    ensureAudioGuidanceRegistered(button.context)
+    ensureAudioGuidanceRegistered()
     return component(AudioGuidanceButtonComponent(button))
 }
 
 @ExperimentalPreviewMapboxNavigationAPI
-private fun ensureAudioGuidanceRegistered(context: Context) {
+private fun ensureAudioGuidanceRegistered() {
     if (MapboxNavigationApp.getObservers(MapboxAudioGuidance::class).isEmpty()) {
-        val audioGuidance = MapboxAudioGuidanceImpl.create(context)
-        MapboxNavigationApp.registerObserver(audioGuidance)
+        MapboxNavigationApp.registerObserver(MapboxAudioGuidance())
     }
 }